nutri-project/requirements.txt
# requirements.txt

Django==5.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2
Pillow==10.3.0
django-cors-headers==4.3.1
